refactor: 칸반 드래그 앤 드롭 로직 분리 및 개선
